@@ -4,14 +4,17 @@ import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVRecord;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.geotools.api.feature.simple.SimpleFeature;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.geom.Point;
 import org.matsim.api.core.v01.Coord;
+import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.TransportMode;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.api.core.v01.network.NetworkWriter;
 import org.matsim.api.core.v01.network.Node;
+import org.matsim.application.prepare.counts.CreateCountsFromBAStData;
 import org.matsim.application.prepare.pt.CreateTransitScheduleFromGtfs;
 import org.matsim.contrib.bicycle.BicycleUtils;
 import org.matsim.contrib.osm.networkReader.LinkProperties;
@@ -20,6 +23,8 @@ import org.matsim.contrib.osm.networkReader.OsmTags;
 import org.matsim.core.controler.OutputDirectoryLogging;
 import org.matsim.core.network.NetworkUtils;
 import org.matsim.core.network.algorithms.MultimodalNetworkCleaner;
+import org.matsim.core.network.algorithms.NetworkSimplifier;
+import org.matsim.core.scenario.ProjectionUtils;
 import org.matsim.core.utils.geometry.CoordinateTransformation;
 import org.matsim.core.utils.geometry.geotools.MGC;
 import org.matsim.core.utils.geometry.transformations.IdentityTransformation;
@@ -29,61 +34,51 @@ import org.matsim.prepare.counts.CombinedCountsWriter;
 import org.matsim.prepare.counts.LongTermCountsCreator;
 import org.matsim.prepare.counts.RawDataVehicleTypes;
 import org.matsim.prepare.counts.ShortTermCountsCreator;
-import org.opengis.feature.simple.SimpleFeature;
 
 import java.io.FileReader;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
+import java.util.function.BiPredicate;
 import java.util.stream.Collectors;
 
 import static org.matsim.prepare.RuhrUtils.*;
 
+/**
+ * Create supply (network, pt, counts) for the scenario.
+ */
 public class CreateSupply {
 
-	public enum NetworkResolution {Low, Medium, High}
-
 	private static final NetworkResolution networkResolution = NetworkResolution.High;
-
 	private static final Path osmData = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/osm/germany-coarse_nordrhein-westfalen-2021-07-09_merged.osm.pbf");
 	private static final Path ruhrShape = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/shp-files/ruhrgebiet_boundary/ruhrgebiet_boundary.shp");
 	private static final Path heightData = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/original-data/2021-05-29_RVR_Grid_10m.tif");
-
 	private static final Path nrwShape = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/shp-files/nrw/dvg2bld_nw.shp");
-
-	private static final Path gtfsData1 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/vrr_20211118_gtfs_vrr_shapes.zip");
-	private static final Path gtfsData2 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/gtfs-nwl-20210215.zip");
-	private static final Path gtfsData3 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/gtfs-schienenfernverkehr-de_2021-08-19.zip");
-
-	private static final String gtfsDataDate1 = "2021-11-17";
-	private static final String gtfsDataDate2 = "2021-02-04";
-	private static final String gtfsDataDate3 = "2021-08-19";
-
-	private static final String gtfsData1Prefix = "vrr";
-	private static final String gtfsData2Prefix = "nwl";
-	private static final String gtfsData3Prefix = "fern";
-
+	//new gtfs file provided by the rvr
+	private static final Path gtfsData1 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/20230106_gtfs_nrw_neue_service_ids_korrektur1.zip");
+	//private static final Path gtfsData1 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/vrr_20211118_gtfs_vrr_shapes.zip");
+	//private static final Path gtfsData2 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/gtfs-nwl-20210215.zip");
+	private static final Path gtfsData2 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/gtfs-schienenfernverkehr-de_2021-08-19.zip");
+	private static final String gtfsDataDate1 = "2023-01-17";
+	//private static final String gtfsDataDate2 = "2021-02-04";
+	private static final String gtfsDataDate2 = "2021-08-19";
+	private static final String gtfsData1Prefix = "nrw";
+	//private static final String gtfsData2Prefix = "nwl";
+	private static final String gtfsData2Prefix = "fern";
 	private static final Path inputShapeNetwork1 = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/original-data/2021-03-05_radwegeverbindungen_VM_Freizeitnetz/2021-03-05_radwegeverbindungen_VM_Freizeitnetz.shp");
 	private static final Path inputShapeNetwork2 = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/original-data/2021-03-05_radwegeverbindungen_VM_Knotenpunktnetz/2021-03-05_radwegeverbindungen_VM_Knotenpunktnetz.shp");
 	private static final Path inputShapeNetwork3 = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/original-data/2021-08-19_radwegeverbindungen_RRWN_Bestandsnetz/2021-08-19_RRWN_Bestandsnetz.shp");
+	private static final Path outputDirPublic = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v2.0/input/");
 	// for now, we will focus on the 'Bestandsnetz'. Once, we are done with calibration, we will also generate the network for the 'Zielnetz' by replacing the previous line with the following.
 	// private static final Path inputShapeNetwork3 = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/original-data/2021-08-19_radwegeverbindungen_RRWN_Bestandsnetz_Zielnetz/2021-08-19_RRWN_Bestandsnetz_Zielnetz.shp");
-
-	private static final Path outputDirPublic = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/input/");
-
-	private static final Path outputDirCounts = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/input/");
+	private static final Path outputDirCounts = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v2.0/input/");
+	private static final Path bastCountsRoot = Paths.get("shared-svn/projects/rvr-metropole-ruhr/data/BASt");
 	private static final Path longTermCountsRoot = Paths.get("shared-svn/projects/matsim-ruhrgebiet/original_data/counts/long_term_counts");
 	private static final Path longTermCountsIdMapping = Paths.get("shared-svn/projects/matsim-ruhrgebiet/original_data/counts/mapmatching/countId-to-nodeId-long-term-counts.csv");
 	private static final Path shortTermCountsRoot = Paths.get("shared-svn/projects/matsim-ruhrgebiet/original_data/counts/short_term_counts");
 	private static final Path shortTermCountsIdMapping = Paths.get("shared-svn/projects/matsim-ruhrgebiet/original_data/counts/mapmatching/countId-to-nodeId-short-term-counts.csv");
-
-	private static final Path parkingShapeFile = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/original-data/Parkraum_20220902/Parkraum.shp");
-
-
+	private static final Path parkingShapeFile = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v2.0/original-data/20240905_Parkraum/20240905_Parkraum.shp");
 	// we use UTM-32 as coordinate system
 	private static final CoordinateTransformation transformation = TransformationFactory.getCoordinateTransformation(TransformationFactory.WGS84, "EPSG:25832");
 	private static final Logger logger = LogManager.getLogger(CreateSupply.class);
@@ -116,31 +111,35 @@ public class CreateSupply {
 		}
 
 		var ruhrGeometries = ShapeFileReader.getAllFeatures(rootDirectory.resolve(ruhrShape).toString()).stream()
-				.map(feature -> (Geometry) feature.getDefaultGeometry())
-				.collect(Collectors.toList());
+			.map(feature -> (Geometry) feature.getDefaultGeometry())
+			.collect(Collectors.toList());
 
 		var nrwGeometries = ShapeFileReader.getAllFeatures(rootDirectory.resolve(nrwShape).toString()).stream()
-				.map(feature -> (Geometry) feature.getDefaultGeometry())
-				.collect(Collectors.toList());
+			.map(feature -> (Geometry) feature.getDefaultGeometry())
+			.collect(Collectors.toList());
 
 		var nodeIdsToKeep = List.of(parseNodeIdsToKeep(rootDirectory.resolve(longTermCountsIdMapping)), parseNodeIdsToKeep(rootDirectory.resolve(shortTermCountsIdMapping))).stream()
-				.flatMap(Collection::stream)
-				.collect(Collectors.toSet());
+			.flatMap(Collection::stream)
+			.collect(Collectors.toSet());
 
 		// ----------------------------------------- Create Network ----------------------------------------------------
 
+		// ruhr area is rather large, with a mixture of urban and more rural areas
+		// based on the paper "Road network free flow speed estimation using microscopic simulation and point-to-point travel times"
+		// the free speed factor was set to a mean of urban, metropolitan and rural areas
 		var networkBuilder = new OsmBicycleReader.Builder()
-				.setCoordinateTransformation(transformation)
-				.setIncludeLinkAtCoordWithHierarchy((coord, level) -> isIncludeLink(coord, level, ruhrGeometries, nrwGeometries))
-				.setPreserveNodeWithId(nodeIdsToKeep::contains)
-				.setAfterLinkCreated((link, tags, direction) -> onLinkCreated(link))
-				.addOverridingLinkProperties(OsmTags.SERVICE, new LinkProperties(10, 1, 10 / 3.6, 100 * 0.25, false)); // set hierarchy level for service roads to 10 to exclude them
+			.setFreeSpeedFactor(0.75)
+			.setCoordinateTransformation(transformation)
+			.setIncludeLinkAtCoordWithHierarchy((coord, level) -> isIncludeLink(coord, level, ruhrGeometries, nrwGeometries))
+			.setPreserveNodeWithId(nodeIdsToKeep::contains)
+			.setAfterLinkCreated((link, tags, direction) -> onLinkCreated(link))
+			.addOverridingLinkProperties(OsmTags.SERVICE, new LinkProperties(10, 1, 10 / 3.6, 100 * 0.25, false)); // set hierarchy level for service roads to 10 to exclude them
 
 		if (networkResolution == NetworkResolution.Low) {
 			// exclude tracks and cycleways
 			networkBuilder
-					.addOverridingLinkProperties(OsmTags.TRACK, new LinkProperties(10, 1, 10 / 3.6, 100 * 0.25, false)) // set hierarchy level to 10 to exclude them
-					.addOverridingLinkProperties(OsmTags.CYCLEWAY, new LinkProperties(10, 1, 10 / 3.6, 100 * 0.25, false)); // set hierarchy level to 10 to exclude them
+				.addOverridingLinkProperties(OsmTags.TRACK, new LinkProperties(10, 1, 10 / 3.6, 100 * 0.25, false)) // set hierarchy level to 10 to exclude them
+				.addOverridingLinkProperties(OsmTags.CYCLEWAY, new LinkProperties(10, 1, 10 / 3.6, 100 * 0.25, false)); // set hierarchy level to 10 to exclude them
 		} else if (networkResolution == NetworkResolution.Medium) {
 			// exclude tracks
 			networkBuilder.addOverridingLinkProperties(OsmTags.TRACK, new LinkProperties(10, 1, 10 / 3.6, 100 * 0.25, false)); // set hierarchy level to 10 to exclude them
@@ -151,24 +150,39 @@ public class CreateSupply {
 		}
 
 		var network = networkBuilder
-				.build()
-				.read(rootDirectory.resolve(osmData));
+			.build()
+			.read(rootDirectory.resolve(osmData));
 
 		//new NetworkWriter(network).write(rootDirectory.resolve(outputDir.resolve("metropole-ruhr-v1.0.network-only-OSM-and-PT_resolution" + networkResolution + ".xml.gz")).toString());
 
 		// ----------------------------- Add bicycles and write network ------------------------------------------------
 
 		Network network1 = new ShpToNetwork().run(rootDirectory.resolve(inputShapeNetwork1));
-		new NetworkWriter(network1).write(outputDir.resolve("metropole-ruhr-v1.4.network-onlyBikeNetwork1.xml.gz").toString());
+		new NetworkWriter(network1).write(outputDir.resolve("metropole-ruhr-v2.0.network-onlyBikeNetwork1.xml.gz").toString());
 		new BikeNetworkMerger(network).mergeBikeHighways(network1);
 
 		Network network2 = new ShpToNetwork().run(rootDirectory.resolve(inputShapeNetwork2));
-		new NetworkWriter(network2).write(outputDir.resolve("metropole-ruhr-v1.4.network-onlyBikeNetwork2.xml.gz").toString());
+		new NetworkWriter(network2).write(outputDir.resolve("metropole-ruhr-v2.0.network-onlyBikeNetwork2.xml.gz").toString());
 		new BikeNetworkMerger(network).mergeBikeHighways(network2);
 
 		Network network3 = new ShpToNetwork().run(rootDirectory.resolve(inputShapeNetwork3));
-		new NetworkWriter(network3).write(outputDir.resolve("metropole-ruhr-v1.4.network-onlyBikeNetwork3.xml.gz").toString());
+		new NetworkWriter(network3).write(outputDir.resolve("metropole-ruhr-v2.0.network-onlyBikeNetwork3.xml.gz").toString());
 		new BikeNetworkMerger(network).mergeBikeHighways(network3);
+
+		var simplifier = new NetworkSimplifier();
+
+		// Original bike network is not merged
+		BiPredicate<Link, Link> isMergeable = (link1, link2) -> !link1.getId().toString().startsWith("bike") && !link2.getId().toString().startsWith("bike");
+
+		simplifier.run(network, isMergeable, NetworkSimplifier.DEFAULT_TRANSFER_ATTRIBUTES_CONSUMER);
+
+		// Clean unused nodes
+		List<Id<Node>> toRemove = network.getNodes().values().stream()
+			.filter(node -> node.getOutLinks().isEmpty() && node.getInLinks().isEmpty())
+			.map(Node::getId)
+			.toList();
+
+		toRemove.forEach(network::removeNode);
 
 		var cleaner = new MultimodalNetworkCleaner(network);
 		cleaner.run(Set.of(TransportMode.car));
@@ -263,48 +277,118 @@ public class CreateSupply {
 				NetworkUtils.setLinkEgressTime(link, TransportMode.bike, 0.0);
 				NetworkUtils.setLinkAccessTime(link, TransportMode.ride, 0.0);
 				NetworkUtils.setLinkEgressTime(link, TransportMode.ride, 0.0);
+				NetworkUtils.setLinkAccessTime(link, "truck8t", 0.0);
+				NetworkUtils.setLinkEgressTime(link, "truck8t", 0.0);
+				NetworkUtils.setLinkAccessTime(link, "truck18t", 0.0);
+				NetworkUtils.setLinkEgressTime(link, "truck18t", 0.0);
+				NetworkUtils.setLinkAccessTime(link, "truck26t", 0.0);
+				NetworkUtils.setLinkEgressTime(link, "truck26t", 0.0);
+				NetworkUtils.setLinkAccessTime(link, "truck40t", 0.0);
+				NetworkUtils.setLinkEgressTime(link, "truck40t", 0.0);
+
 			}
 		}
 
-		String networkOut = outputDir.resolve("metropole-ruhr-v1.4.network_resolution" + networkResolution + ".xml.gz").toString();
-		new NetworkWriter(network).write(networkOut);
+		String networkOut = outputDir.resolve("metropole-ruhr-v2.0.network_resolution" + networkResolution + ".xml.gz").toString();
+
+		fixLinks(network);
+		addCommercialTrafficModes(network);
+
+		ProjectionUtils.putCRS(network, "EPSG:25832");
+
+		NetworkUtils.writeNetwork(network, networkOut);
 
 		// --------------------------------------- Create Pt -----------------------------------------------------------
 
+		String outputName = "metropole-ruhr-v2.0";
 		new CreateTransitScheduleFromGtfs().execute(
-				rootDirectory.resolve(gtfsData1).toString(), rootDirectory.resolve(gtfsData2).toString(), rootDirectory.resolve(gtfsData3).toString(),
-				"--date", gtfsDataDate1, gtfsDataDate2, gtfsDataDate3,
-				"--prefix", gtfsData1Prefix + "," + gtfsData2Prefix + "," + gtfsData3Prefix,
-				"--target-crs", "EPSG:25832",
-				"--network", networkOut,
-				"--output", outputDir.toString(),
-				"--name", "metropole-ruhr-v1.4"
+			rootDirectory.resolve(gtfsData1).toString(), rootDirectory.resolve(gtfsData2).toString(),
+			"--date", gtfsDataDate1, gtfsDataDate2,
+			"--prefix", gtfsData1Prefix + "," + gtfsData2Prefix,
+			"--target-crs", "EPSG:25832",
+			"--network", networkOut,
+			"--output", outputDir.toString(),
+			"--copy-late-early=true",
+			"--validate=true",
+			"--name", outputName
+		);
+
+		// --------------------------------------------------------------------
+
+		new TagTransitSchedule().execute(
+			"--input", outputDir + "/" + outputName + "-transitSchedule.xml.gz",
+			"--shp", rootDirectory.resolve(ruhrShape).toString(),
+			"--output", outputDir + "/" + outputName + "-transitSchedule.xml.gz"
 		);
 
 		// --------------------------------------- Create Counts -------------------------------------------------------
 
 		var longTermCounts = new LongTermCountsCreator.Builder()
-				.setLoggingFolder(outputDirForCounts + "/")
-				.withNetwork(network)
-				.withRootDir(rootDirectory.resolve(longTermCountsRoot).toString())
-				.withIdMapping(rootDirectory.resolve(longTermCountsIdMapping).toString())
-				.withStationIdsToOmit(5002L, 50025L)
-				.useCountsWithinGeometry(rootDirectory.resolve(ruhrShape).toString())
-				.build()
-				.run();
+			.setLoggingFolder(outputDirForCounts + "/")
+			.withNetwork(network)
+			.withRootDir(rootDirectory.resolve(longTermCountsRoot).toString())
+			.withIdMapping(rootDirectory.resolve(longTermCountsIdMapping).toString())
+			.withStationIdsToOmit(5002L, 50025L)
+			.useCountsWithinGeometry(rootDirectory.resolve(ruhrShape).toString())
+			.build()
+			.run();
 
 		var shortTermCounts = new ShortTermCountsCreator.Builder()
-				.setLoggingFolder(outputDirForCounts + "/")
-				.withNetwork(network)
-				.withRootDir(rootDirectory.resolve(shortTermCountsRoot).toString())
-				.withIdMapping(rootDirectory.resolve(shortTermCountsIdMapping).toString())
-				.withStationIdsToOmit(5002L, 50025L)
-				.useCountsWithinGeometry(rootDirectory.resolve(ruhrShape).toString())
-				.build()
-				.run();
+			.setLoggingFolder(outputDirForCounts + "/")
+			.withNetwork(network)
+			.withRootDir(rootDirectory.resolve(shortTermCountsRoot).toString())
+			.withIdMapping(rootDirectory.resolve(shortTermCountsIdMapping).toString())
+			.withStationIdsToOmit(5002L, 50025L)
+			.useCountsWithinGeometry(rootDirectory.resolve(ruhrShape).toString())
+			.build()
+			.run();
 
-		CombinedCountsWriter.writeCounts(outputDirForCounts.resolve("metropole-ruhr-v1.4.counts.xml.gz"),
-				longTermCounts.get(RawDataVehicleTypes.Pkw.toString()), shortTermCounts.get(RawDataVehicleTypes.Pkw.toString()));
+
+		new CreateCountsFromBAStData().execute(
+			"--network", networkOut,
+			"--year", "2022",
+			"--shp", rootDirectory.resolve(ruhrShape).toString(),
+			"--shp-crs", "EPSG:25832",
+			"--counts-mapping", rootDirectory.resolve(bastCountsRoot).resolve("bast-mapping.csv").toString(),
+			"--motorway-data", rootDirectory.resolve(bastCountsRoot).resolve("2022_A_S.zip").toString(),
+			"--primary-data", rootDirectory.resolve(bastCountsRoot).resolve("2022_B_S.zip").toString(),
+			"--station-data", rootDirectory.resolve(bastCountsRoot).resolve("Jawe2022.csv").toString(),
+			"--output", rootDirectory.resolve(outputDirPublic).resolve("metropole-ruhr-v2.0.counts.xml.gz").toString()
+		);
+
+
+		// Based on 2015 count data, won't be used in the new model
+		CombinedCountsWriter.writeCounts(outputDirForCounts.resolve("metropole-ruhr-v2.0.counts-old.xml.gz"),
+			longTermCounts.get(RawDataVehicleTypes.Pkw.toString()), shortTermCounts.get(RawDataVehicleTypes.Pkw.toString()));
+	}
+
+	/**
+	 * Fix links with speed 0
+	 */
+	private void fixLinks(Network network) {
+		for (Link link : network.getLinks().values()) {
+			// set freespeed to 10 km/h if it is 0
+			if (link.getFreespeed() == 0) {
+				link.setFreespeed(10 / 3.6);
+			}
+
+			if (link.getLength() == 0) {
+				link.setLength(1);
+			}
+		}
+	}
+
+	/**
+	 * Add network modes required by commercial traffic
+	 */
+	private void addCommercialTrafficModes(Network network) {
+		for (Link link : network.getLinks().values()) {
+			if (link.getAllowedModes().contains(TransportMode.car)) {
+				var modes = new HashSet<>(link.getAllowedModes());
+				modes.addAll(Set.of("freight", "truck8t", "truck18t", "truck26t", "truck40t"));
+				link.setAllowedModes(modes);
+			}
+		}
 	}
 
 	private boolean isIncludeLink(Coord coord, int level, Collection<Geometry> ruhrGeometries, List<Geometry> nrwGeometries) {
@@ -340,9 +424,9 @@ public class CreateSupply {
 			link.setAllowedModes(modes);
 		}
 
-		// all regular bike links should have an infrastructure speed factor of 0.5
+		// all regular bike links should have an infrastructure speed factor of 0.64 according to the rvr
 		if (link.getAllowedModes().contains(TransportMode.bike)) {
-			link.getAttributes().putAttribute(BicycleUtils.BICYCLE_INFRASTRUCTURE_SPEED_FACTOR, 0.5);
+			link.getAttributes().putAttribute(BicycleUtils.BICYCLE_INFRASTRUCTURE_SPEED_FACTOR, 0.64);
 		}
 	}
 
@@ -364,5 +448,7 @@ public class CreateSupply {
 		}
 		return result;
 	}
+
+	public enum NetworkResolution {Low, Medium, High}
 
 }
