@@ -19,15 +19,15 @@
 
 package org.matsim.gui;
 
-import org.matsim.run.RunMetropoleRuhrScenario;
+import org.matsim.run.MetropoleRuhrScenario;
 import org.matsim.run.gui.Gui;
 
 public class MATSimGUI {
 
 	public static void main(String[] args) {
-		Gui.show("MATSim GUI from example project", RunMetropoleRuhrScenario.class );
+		Gui.show("MATSim GUI from example project", MetropoleRuhrScenario.class );
 	}
 
-	//	The jar file is generated by "mvn package".  Note that the pom.xml refers at some point to this class here. 
+	//	The jar file is generated by "mvn package".  Note that the pom.xml refers at some point to this class here.
 
 }
