@@ -8,8 +8,8 @@ public class DeleteRoutesFromPlans {
 
     public static void main(String[] args) throws Exception {
         String[] argsForRemoveRoutesFromPlans = new String[]{
-                "--plans=../../shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/input/metropole-ruhr-v1.4-3pct.plans.xml.gz",
-                "--output=../../shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/input/metropole-ruhr-v1.4-3pct.plans-withoutRoutes.xml.gz",
+                "--plans=../../../public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v2.0/input/metropole-ruhr-v2.0-3pct.plans-commercial.xml.gz",
+                "--output=../../../public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v2.0/input/no-routes/metropole-ruhr-v2.0-3pct.plans-commercial.xml.gz",
                 "--remove-routes=true",
                 "--remove-unselected-plans=true",
                 "--remove-activity-location=true"
